PoC demonstrates how to use PageFactory and SharedWebDriver
