+++ jeta-master/.gitignore
@@ -3,3 +3,5 @@
 *.iml
 /.idea
 /tools
+/config/creds.yml
+/config/
