+++ jeta-master/pom.xml
@@ -40,13 +40,24 @@
         </dependency>
 
         <!-- https://mvnrepository.com/artifact/info.cukes/cucumber-java -->
-        <dependency>
-            <groupId>info.cukes</groupId>
-            <artifactId>cucumber-java</artifactId>
-            <version>1.2.5</version>
-        </dependency>
+         <dependency>
+             <groupId>info.cukes</groupId>
+             <artifactId>cucumber-java</artifactId>
+             <version>1.2.5</version>
+         </dependency>
 
+         <dependency>
+             <groupId>info.cukes</groupId>
+             <artifactId>cucumber-picocontainer</artifactId>
+             <version>1.2.5</version>
+         </dependency>
 
-    </dependencies>
+          <dependency>
+              <groupId>org.picocontainer</groupId>
+              <artifactId>picocontainer</artifactId>
+              <version>2.15</version>
+          </dependency>
 
-</project>
+     </dependencies>
+
+ </project>
